requirements.txt
asgiref==3.8.1
Django==4.2
django-tinymce==4.1.0
pillow==11.1.0
PyMySQL==1.1.1
sqlparse==0.5.3
